Refactor WordPress demo to use a local sqlite database
This way the demo is self-contained and doesn't need to worry
about an external MySQL database and credentials.

This also fits well with the Now paradigm of immutable deployments,
where any modification to the local sqlite database is embedded
into the deployment and the data will stay "static" for this deployment
even as the database evolves with more content in the future.

TODO: Update root Readme with build size info, and add a Readme to
this example with instructions on updating the db content locally.