Provide info about Deep API Account in Readme.md
Cleaned up the readme a bit to save anyone time who wants to actually try to get this working.  This saves having to search the code to find the  URL ot the DeepAI service.

Also added a bit of grouping of the steps with subheads (e.g. "Store the Tokens and Keys as now secrets")